+++ examples/assets/rendering-test.js
@@ -144,7 +144,7 @@ function initDesigner(testCaseName) {
 		},
 
 		editors: {
-			globalEditorProvider: () => {
+			rootEditorProvider: () => {
 				const editor = document.createElement('div');
 				editor.innerText = 'Please select any step.';
 				return editor;